Fixes issue with useCheckoutPricing currency
We discovered an issue with useCheckoutPricing not
raising an error when passing a currency that isn't accepted on a plan.

To fix the issue, we needed to rearrange the call order inside
`useCheckoutPricing` to call subscriptions first, then adjustments,
then everything else.

To add an assertion for this, I also added `jest-extended` for its
`toHaveBeenCalledBefore` matcher.